Use language only to identify Spanish translation files
This  allows the  Spanish UI  translation to  work, for  now, for  all
Spanish speaking locales.  If necessary we  can make it es-ES if other
translators feel it is not a good base for their Spanish variant. OTOH
if they just need to l10n a  few strings then, say for Argentina, then
they can do  that in an wsjtx_es_AR.ts and  untranslated strings there
will fall back to the  ones in wsjtx_es.ts automatically. This happens
because of teh way the application loads multiple translation files in
an order suitable for that to happen.