Environment variables to set audio buffer sizes and fix Windows Rx timing
The two environment variables:

    WSJT_RX_AUDIO_BUFFER_FRAMES
    WSJT_TX_AUDIO_BUFFER_FRAMES

each can  be defined to  an integer number which  will be used  as the
suggested audio  buffer size for  Rx and Tx respectively.  Not setting
the variable  or setting  it to  zero or less  will cause  the default
buffer size to be used, which should be a good choice for most, if not
all, systems.