Generate a tune up tone when WSPR band hopping for Tx periods
Previously a tune up  tone was not generated if the  first period on a
new band was  going to be a Tx  period, this is no longer  so and thus
allows low (or high) power tune tones as set up by the user. Note that
tune up tones are  still not sent for bands flagged as  Rx Only in the
band hopping  schedule, this  is to protect  active aerials  and avoid
unintended transmissions on unlicensed bands.