Remove a, hopefully, unnecessary unconditional CAT frequency set
this frequency  set was  causing CAT command  rejections on  some Icom
rigs when  using WSPR  auto tune-up with  the rig's  built-in auto-ATU
anabled.    The  issue   relates  to   a  PTT-set,   tune,  PTT-reset,
set-frequency sequence where  the rig's ATU is still  tuning while the
PTT is reset.  Since this should  only happen in WSPR mode where split
operating is not used it should resolve the issue.